+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 - `Call` objects for the same `cId` will reference the same memory instance. [#404](https://github.com/GetStream/stream-video-swift/pull/404)
+- `CallKitService.callEnded` now accepts the cId of the call to end. [#406](https://github.com/GetStream/stream-video-swift/pull/406)
+- `CallKitService.State` has been deprecated. [#406](https://github.com/GetStream/stream-video-swift/pull/406)
 
 ### 🐞 Fixed
 - Video tracks for anonymous users not displaying